+++ lib/wid_design_system.dart
@@ -3,10 +3,11 @@ library wid_design_system;
 
 export 'src/providers/wid_theme_provider.dart';
 export 'src/theme/palette.dart';
-export 'src/theme/text.dart';
+export 'src/theme/text_styles.dart';
 export 'src/theme/theme.dart';
 export 'src/utils/dimensions.dart';
 export 'src/wid_design_system.dart';
 export 'src/widgets/wid_button.dart';
 export 'src/widgets/wid_spacer.dart';
 export 'src/widgets/wid_text.dart';
+export 'src/widgets/wid_touchable.dart';