+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.0-beta.1
+
+- feat: Added custom button with 4 variants 🎉
+- feat: Added several theme implementations for several widgets 🎉
+- feat: Added WidTapToHideKeyboard widget 🎉
+- feat: Added custom touchable widget 🎉
+- feat: Exposing and providing fonts 🎉
+
 # 0.1.0-beta.1
 
 - feat: initial commit 🎉